Removed files in phatbooks_text_session from list of sources for phatbooks_test executable - as they are not used in the current tests at all.
Introduced "BString" as a typedef for "whatever the string type it
is that is used at the business layer". Allows the type to be
changed more easily later if we want to. Introduced appropriate
functions to convert back and forth between BString and std::string
(at the text user interface layer) and BString and wxString (at
the GUI layer).

BString is for the time being defined as std::string. Note that if defined as
wxString, a strange bug arises that involves garbage being
written to database. Still need to determine cause.

Various play code (now commented out) inserted into test main function in
an attempt to pin down Valgrind errors that have now also been discovered.
The cause of these errors (and whether they are genuine) has yet to be
determined.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1053 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98