Removed #includes from test/test.cpp that are no longer required.
Rearranged various typedefs of, and references to, "Id" types (e.g.
AccountImpl::Id, Journal::Id...) to improve their
modularity in respect of the individual classes to which
they pertain. This improves the robustness of the design - even though
in practice these Id types are all just synonyms for sqloxx::Id.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1070 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98