Fixed error in assert in Session constructor.
Provided single macro (PHATBOOKS_ON_WINDOWS) to state whether we're
on Windows.

Enabling program to be run without having to pass a filepath as a command
line argument. In this case the user may open a database file from within
the Session. (WORK IN PROGRESS.)

This is a NONCOMPILING commit.



git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1092 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98