Abolised CurrencyManager class, and instead just exposed currency creation functionality via a free function. It wasn't necessary to have a class here.
Renamed make_currencies.inc to make_currencies_inc.hpp so that CMake
will pick up changes when building.

Changed SetupWizard in light of new API for currency creation. Used
SetClientObject and GetClientObject to associate a Commodity with
each item in the wxComboBox via which the user selects their currency -
rather than having to maintain a map from wxStrings to Commodities
as previously.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1377 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98