Actually implemented destination_account_types(...), which I forgot to do prior to last commit.
Wrote tests for source_account_types(...) and destination_account_types(...).
All tests pass.

Fixed error in test_draft_journal_repeater_description, due to next-date being
in the past. This was an inconsequential error but caused the test to fail.
Now made the next-dates in this test well into the future so this failure
won't happen again - for many generations.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1588 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98