Got rid of m_actual_vs_budget_ctrl from TransactionCtrl; it is now implicit in the TransactionType, whether it is an actual or budget transaction.
Instead of a DecimalTextCtrl for each Entry row in TransactionCtrl,
we now have only one DecimalTextCtrl - for the "primary amount" of the
transaction - and we have "Split..." buttons (as yet inoperable) where
each Entry row's DecimalTextCtrl used to be.

Fixed TransactionType::post_journal() so that it now works with the new
TransactionCtrl structure.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1598 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98