AccountDialog can now be used successfully to create new Accounts.
A couple of issues remain to be resolved, namely
that balance sheet AccountListCtrl is not being updated for
newly created balance sheet Account, and the signs for opening
balances are sometimes the reverse of what is expected. I have
documented these as TODOs in account_dialog.hpp.

Also spacing and sizing of AccountDialog sub-widgets needs a
bit of work.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1681 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98