Merged SetupWizard::BalanceSheetAccountPage and SetupWizard::PLAccountPage into SetupWizard::AccountPage - now it's just this one class. Created a second AccountPage in SetupWizard for configuring initial P&L Accounts. Various refactoring in SetupWizard and MultiAccountPanel.
Still need to arrange things so that the P&L Account Page in the SetupWizard
actually causes appropriate P&L Accounts to be created.

Still need to fix a sizing issue with scrolling in AccountPage.

Still need to provide more explanation etc.. in AccountPage to assist
user with the task of allocating opening balances and initial budget
allocations.

Various temporary debugging statements currently distributed throughout this
and related code.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1900 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98