Continued working on MultiAccountPanel and SetupWizard.
Fixed an issue with MultiAccountPanel where the signs were the wrong
way round for the opening balances when m_account_super_type was
account_super_type::pl.

Fixed SetupWizard so that it creates not only balance sheet Accounts but
also P&L Accounts on the basis of what the user has entered into the
MultiAccountPanels.




git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@1902 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98