Updated some TODOs.
Changed TransactionCtrl so that in order to construct a TransactionCtrl
for the purpose of creating a new Journal (as opposed to editing an
existing one), you pass it a ProtoJournal, rather than passing
it vectors of Accounts etc.. as previously. This is a simpler and much
less confusing interface than previously.

Updated code in TopPanel in light of the new interface with TransactionCtrl.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2023 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98