Abolished function commonest_actual_transaction_type_since(...), which it turns out there was not required at all for its intended purpose.
Made it so when the application initially opens, an Account is pre-selected
in each of TopPanel::m_bs_account_list and TopPanel::m_pl_account_list,
based on which has been used most often in actual OrdinaryJournals. This
in turn means that the TransactionCtrl is pre-populated with these
Accounts, being default selections which the user has a reasonable chance
of wanting to use.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2024 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98