Renamed TopPanel::m_transaction_panel to "m_entry_list_panel". The old name was misleading; "m_entry_list_panel" reflects the fact that it points to an EntryListPanel.
Enabled user to toggle whether Accounts are visible or hidden in the
AccountListCtrls, via a checkbox in AccountDialog. Also, enabled user to
toggle, via "View" menu at top of Frame, whether AccountListCtrls display
or hide the "hidden" Accounts. (Note it is currently only in AccountListCtrl
that the hidden / visible status of an Account is relevant to what is
displayed in the GUI.)


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2029 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98