Abolished functions finformat_bstring(...) and finformat_bstring_nopad(...) which were not being used.
Renamed PhraseFlags to StringFlags, to reflect a broadening
of its uses, to include flags for numeric formatting and parsing
as well as for natural language phrase formation.

Extended StringFlags to incorporate information about formatting
and parsing of jewel::Decimal.

Changed functions in "finformat.hpp" to make use of StringFlags, and
made corresponding changes at call sites.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2044 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98