Made it so that, in functions where we are traversing an AccountListCtrl extracting Accounts one by one, we do so using the Id rather than the Account name. This is more efficient, and also means that if the Account name has changed, the extraction attempt will still succeed. (This was causing a crash when updating for an Account where we just edited the name.)
Made it so that, when an Account is updated, it does not
cause the TransactionCtrl to be cleared. Rather, the
AccountCtrls, within the EntryGroupCtrls, within the
TransactionCtrl, are simply updated to reflect the new
details of the Account.

Made it so that the events PHATBOOKS_JOURNAL_CREATED_EVENT and
PHATBOOKS_JOURNAL_EDITED_EVENT, when intercepted by TopPanel, do
NOT cause the TransactionCtrl to be cleared. This had caused
an obscure bug (involving the created of opening balance Journals
from AccountDialog causing TransactionCtrl to inappropriately clear)
and had the potential to cause more. Rather, the TransactionCtrl itself
now directly controls when it is cleared in response to the "Save",
"Delete" or "Clear" buttons being pressed within the TransactionCtrl.

In AccountDialog, discovered that an new opening balance Journal
was being created and saved every time the user clicked the OK button -
even if there was no change to the opening balance. Fixed this so that
we don't save a new opening balance Journal if the opening balance
hasn't changed.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2049 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98