Made it so that EntryListPanel and ReportPanel do not generate their EntryListCtrl/Report until the user first presses the "Run" button. This speeds up application startup. Furthermore, for some mysterious reason, these reports weren't physically displaying until the user first clicked "Run" anyway - so we haven't actually lost anything by explicitly delaying the construction of the corresponding objects till that first click.
Changed the labels on the "Run" and "Refresh" buttons in these windows, so
that in EntryListPanel, they always say "Run", whereas in Report, it
says "Run" when first created, and then changes to "Refresh" on first
click. This is supposed to hint to the user that the Report will not
refresh at all without being clicked; whereas the EntryListPanel
will automatically refresh for new Entries etc.. but will not be "run" for new
parameters (Account and date range) until "Run" is clicked.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2217 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98