Improved some API documentation.
Deleted some obsolete TODOs.

Added priority specifications to some TODOs.

Provided optimisation in BalanceSheetReport::refresh_map() for
the special case in which opening and closing date controls are
left blank: in this case we can just use the opening balance and current
balance of each Account for their opening and closing balances as
displayed in the report.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2459 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98