Made Commodity::precision() and set_precision(...) use jewel::Decimal::places_type, instead of int as previously.
Made it so that user can select a different currency precision
on SetupWizard::FilepathPage, from the "official" precision for their
currency, although the official precision is shown by default. This is
useful for currencies like Japanese Yen where the official precision (in
the case of Yen, 2 decimal places) is more precise than the precision
that is of practical use for day-to-day transactions.

Added a TODO re. "duplicate Account name" error message being incorrectly
invoked in certain scenarios in SetupWizard.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2525 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98