Changed AccountListCtrl::default_account() to return a sqloxx::Handle<Account>, instead of a boost::optional<sqloxx::Handle<Account> >.
Made it so that AccountListCtrl::default_account() will return a (Handle to a)
hidden Account, in the event that there are no visible Accounts to return. This
prevents a possible crash if the user makes all Accounts hidden.

Explicitly documented the fact that when TransactionCtrl is used to create a new
Journal, this relies on there being at least 2 Entries in the ProtoJournal that
is passed to the relevant TransactionCtrl constructor. Also made explicit
comment to similar effect in body of TopPanel::configure_transaction_ctrl().

Some other minor edits.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2535 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98