Abolished unused typedef, BasicDecimalFormatFlags.
In finformat_wx(...), edited the implementation to bring it into line with the
API documentation, so that now when formatting Decimal zero, we do NOT add
extra spaces as a "hack" to make the dash line up with the decimal point
when presented in a column in non-fixed-width fonts. This hack was only
ever going to work with certain fonts anyway; and the aesthetic benefit is
arguably negligible. Better to have clean explainable behaviour for the
function.

Wrote tests for finformat_wx(...). Tests pass.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2540 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98