Caused UnbalanceJournalException to derive from InvalidJournalException.
Created JournalOverflowException, to be thrown when posting a given
PersistentJournal would cause overflow or precision loss in Account balances.
Caused this to be thrown rather than allowing such a PersistentJournal to
be posted.

Made it so that InvalidJournalException is thrown in case an attempt is made
to save a budget Journal that contains an Entry with an Account that is not of
AccoutSuperType::pl.

Made it so that InvalidBudgetItemException is thrown in case an attempt is made
to save a BudgetItem for which the Account is not of AccountSuperType::pl.

Fixed some stuff in Account tests to avoid throwing due to BudgetItems
with Accounts not of AccountSuperType::pl.

Updated some TODOs.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2550 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98