Abolished Application class and rolled its functions into the App class.
In CMakeLists.txt, rolled "common_sources" and "gui_sources" into a single
set of sources, all of which is compiled into the "phatbooks_common"
intermediate static library. This library is then linked to both by the
test executable and by the main Phatbooks executable. This is a simpler
approach now that "app.cpp" is no longer being considered one of the "gui"
sources.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2555 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98