Minor comment and documentation edits.
Minor formatting edits.

Minor tidy-ups in "app.cpp".

In App class, started work on creation of a backup file at the start of each
session, and removal of that file at the end of the session.

Added to Account tests, to exercise BalanceCache::refresh_all() (not just
BalanceCache::refresh_targetted(...)). Tests all pass.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2559 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98