Added some logging statements here and there.
Abolished MultiAccountPanel::SpecialDecimalTextCtrl. Now just using ordinary
DecimalTextCtrl in MultiAccountPanel, and overriding
TransferDataToWindow in MultiAccountPanel to provide the desired behaviour.

Some reworking in BudgetPanel implementation. BudgetPanel is now such that if
the user changes a BudgetItemComponent amount in such a way that it causes the
setting of the summary text to throw a DecimalException, then the contents of
the BudgetPanel will revert to how they were before the exception-causing
change, and a message will be displayed to the user informing them that
the amount or frequency entered cannot be safely processed.

Changed DecimalTextCtrl::do_on_kill_focus(...) to ensure that if parent
window Validate() and TransferDataToWindow() return false, then the
DecimalTextCtrl is reset to its original value.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2585 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98