Made it so that if Repeaters fire on startup of application, the user is informed of which recurring transactions (i.e. DraftJournals) have been automatically recorded, by means of a wxMessageBox displayed on startup.
If any Repeaters are not fired due to JournalOverflowException being
thrown, the exception is caught, and the user is informed of the
missed transactions via the same wxMessageBox. However, if the
budget instrument Journal can't be posted, this is NOT currently
captured by this mechanism in any special way - it is advised to
the user as if it were a normal missed DraftJournal (which isn't good
enough).


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2595 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98