Made some refinements to Frame::report_repeater_firing_results(...). Repeater firings that fail are now reported in a separate wxMessage box with an "exclamation" icon and the title "WARNING" (instead of just the title "Message").
Also made it so that if budget instrument does not fire due to
JournalOverflowException, the user is informed of this
fact via a special wxMessageBox, which is separate again from
the wxMessageBox that reports Repeater failings for "normal"
DraftJournals.

Also, the user will now NOT be informed if the budget instrument
just posts normally (this would be tedious, since it is posted
every day, and it would also confuse the user, to whom the daily
budgets are not normally manifested as a repeating DraftJournal
as such).


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2596 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98