Added "backup.hpp" and "backup.cpp" to version control (having forgot to do so in earlier commit).
Renamed "LICENSE.txt" to "LICENSE", consistent with my other projects.

Applied the Apache License, Version 2.0 to the Phatbooks project, with the
the standard copyright notice block for that license now appearing at the top
of all ".hpp" and ".cpp" files in the project. (Still need to put it into
other kinds of files.)


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2686 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98