Corrected a comment in "test_driver.tcl".
Further worked on "CMakeLists.txt". Now supports binary installation from
sources on Windows as well as just on Unix. Also now on Unix, "make package"
will cause the a source tarball to be produces, whereas on Windows, it
will cause an NSIS installer to be created.

Removed "make_currencies_inc.hpp" from version control, since it is now
generated as part of the build process.


git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2701 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98